Factor idManagement into its own module
